feat: 상품등록/상품수정 화면 공통구현을 위한 ProductEditingView 구현
